training/requirements.txt
# BFMC Vision Training Requirements
# Tested with Python 3.10+ and CUDA 12.x

# Core ML framework
ultralytics>=8.4.7
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0,<2.0  # Pin to 1.x for matplotlib compatibility
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Configuration
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0

# Export formats (optional)
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
# openvino>=2023.0  # For OpenVINO export
# tensorrt>=8.6     # For TensorRT export (installed via NVIDIA)
